DatePipe fix
The DatePipe was throwing exceptions for dates that didn't fit the ISO format. This resulted in console errors and a failure to display dates when they are in a collection - the date is basically formatted twice, except the second time it throws an exception because it receives the already-formatted date.

This will now attempt to parse dates it receives that don't correspond to the ISO format and, when it encounters one, format it appropriately.