Continued work
Got rid of the Profile being passed around - it's not a good solution and there's an outstanding question about whether or not it matters that a task is assigned to the person doing the case progression. TBD.

Added the logic and functionality for determining whether or not the tasks associated with the case are appropriate (and should be closed) or whether there are too many and a warning should be shown.

Also added a bunch of unit tests around that logic/functionality.