Validation fixes
Hidden fields are now disabled to fix the validation on the containing group/array.

Also fixed up the mechanism for returning to the Previous page, which was losing the hidden states of the controls.

TODO: The data being submitted to the API can include nested labels, which wasn't happening previously, so this needs sorting out.