FixedLists
Added some pretty nasty looking code to get FixedLists working in the WriteCollectionFieldComponent. It's very hacky but it produces the right output for the backend and allows the case to progress to the next wizard step.

It _might_ break something else with collections so will need to be tested... but I don't yet know where there are other collections to try it out on.