Fix for complex fields within collections
The formArray should not be passed as the parent as it overwrites the structure set up in buildCaseField so, instead, the "value" group should be passed as the parent. Correspondingly, when a complex group identifies that it's within a collection, it should use the container it's been passed rather than create a complexGroup within it.