Update write-complex-field.component.ts
We need to know not only if a complex field is within a collection but ALSO that it's at the top level within a collection - only then do we reuse the parent as the complexGroup.

If it's a complex field within a complex field within a collection, it should create a new complexField for itself and disregard the parent.