Got address fields showing and hiding
Changed the mechanism for iterating through the fields within a WriteComplexFieldComponent so that it's no longer binding to a pipe and a buildField() method. This is in line with the mechanism introduced into the WriteCollectionFieldComponent.

In addition, the new mechanism ensures that the CaseField is a proper CaseField rather than a simply JSON object, which is something we were seeing.