EUI-3038: Fix for superfluous null value submission
Ensure that any hidden fields which do not have a value are not given a `null` value on form submission. Furthermore, any hidden fields with `retain_hidden_value` set to `true` are removed from the form submission, since these fields' values should not be updated. This fix applies to all field types, including Complex and Collection types, and combinations thereof.