EUI-3267
Initial work to get the grey bars showing for fields with a show_condition that evaluates as true - i.e., they _could_ be hidden but they're showing.

What's not handled in this commit is _not_ showing grey bars in some specific circumstances - they're showing in this implementation but they don't show on AAT so they need special handling. Working on it.