EUI-3490
The previous implementation used URLSearchParams, which automatically ignored null properties. Switching to HttpParams introduced an issue where such nulls are not ignored, which had the consequence of sending "&state=null" when the user selected "Any" in the State filter.

Rather than revert back to the deprecated URLSearchParams or refactor additional areas for the now built-in URLSearchParams, I'm just null-checking and making sure they don't get into the query string.