EUI-3494, EUI-3495
Fix for case creation and a clean-up of some code changes that were made when moving from Http to HttpClient. This was also a problem for the completion of events on the case.

In the midst of this, tidy up some additional points where changes had been made but code had been left in that was redundant (e.g., map(response => response) is a dead operation) or confusing (in the same example, it wasn't actually the response, it was the response's _body_.