EUI-3452
Sizeable refactor to get conditional show/hide working correctly with on the Check your answers page.

The biggest change is that the new ReadFieldsFilterPipe takes a parameter for the top-level FormGroup, which is something that now must be passed down through the layers, alongside the existing [formGroup] input parameter.

This is needed to support conditional show/hides that depend on a "cousin" field - i.e., a field that is a child of the current field's parent's sibling. There are setups with this in place and the prior mechanism simply didn't work because a field only had "access" to its parent and its siblings.

There were also peculiarities with collections that are being handled in the same way as they're being handled for writeable fields - by passing down an idPrefix parameter, which facilitates generation of a full path to the field being evaluated, which in turn allows a full path for the enabling fields to be generated so they can be correctly assessed.