EUI-3452
Fix for table rows still showing for hidden fields, resulting in weird extra grey lines (borders).

Also fixed an issue with the submit button being disabled on the Check your answers page. This was dependent on the validity of a form in a readonly state but didn't account for hidden fields (which are not disabled in the readonly version). The check on the form validity has been removed as it's superfluous. As part of this, also changed it from a function binding to a property binding to improve performance and prevent repeated evaluation.