EUI-3452: Check your answers (#732)
* add cddPageFields filter and ccdTabsFields filter pipe

* remove the pre tag in case history component

* EUI-3452

Sizeable refactor to get conditional show/hide working correctly with on the Check your answers page.

The biggest change is that the new ReadFieldsFilterPipe takes a parameter for the top-level FormGroup, which is something that now must be passed down through the layers, alongside the existing [formGroup] input parameter.

This is needed to support conditional show/hides that depend on a "cousin" field - i.e., a field that is a child of the current field's parent's sibling. There are setups with this in place and the prior mechanism simply didn't work because a field only had "access" to its parent and its siblings.

There were also peculiarities with collections that are being handled in the same way as they're being handled for writeable fields - by passing down an idPrefix parameter, which facilitates generation of a full path to the field being evaluated, which in turn allows a full path for the enabling fields to be generated so they can be correctly assessed.

* EUI-3452

Fixed a bunch of unit tests and, in the process, identified a place where things weren't being hidden in `case-history.component.html`. Also sorted out a lint issue with a line that was too long.

* Update package.json

Version for tagging.

* Update package.json

Version update. Last one wasn't valid.

* EUI-3452

Fix for labels showing twice on the Check your answers page.

* Update RELEASE-NOTES.md

* Update package.json

* Fix for isSolicitor issue

Found that there were intermittent console errors when trying to access `this.user.idam.roles` from within `isSolicitor()` in the Profile class.

I've added some defensive code around it and some unit tests to ensure that nulls are gracefully handled.

* EUI-3452

Fix for table rows still showing for hidden fields, resulting in weird extra grey lines (borders).

Also fixed an issue with the submit button being disabled on the Check your answers page. This was dependent on the validity of a form in a readonly state but didn't account for hidden fields (which are not disabled in the readonly version). The check on the form validity has been removed as it's superfluous. As part of this, also changed it from a function binding to a property binding to improve performance and prevent repeated evaluation.

* New version for a tagged release

* v2.74.3-eta-check-your-answers

Tagging doesn't seem to be working from the command line... so trying for yet another version!

Co-authored-by: kuda <ku-da@hotmail.co.uk>
Co-authored-by: Paul Graham <paul@smartcasual.io>