EUI-3708
Fixed an issue with some fields not showing on the Check your answers page because of a disparity between the CaseField values on the form and on each page. I think there's some cloning going on but the ones on the form are the ones we should be using anyway so that's now what's being used.