EUI-2504
Navigating away from a page that's being edited now displays a modal to inform the user that their unsaved changes will be lost. The message when navigating within the app is different to the one when reloading, closing, or changing the address bar but that's because modern browsers no longer support custom messages.

There are no unit tests in this commit. They are yet to be added but this is functional at the moment.