EUI-3797
Fixed an issue with the Payments History tab having its content missing. The logic to show and hide the contents of tabs depends on there actually being something to show. For this particular tab, the `value` is set to null, which is why the contents are hidden... but this one needs to be treated differently due to how it behaves - it renders a component that goes off and gets its data on its own.