EUI-3825
Fixed an issue where fields on one tab weren't shown if they depended on fields from another tab. This was visible on the FTPA tab within the IA jurisdiction but was actually a generic issue. The solution was to essentially the same as one that was put in place for the Check your answers page.

Some of the changes are simply cleaning up the code within the CaseViewerComponent - something we're doing as we go.