Update loading.service.spec.ts
Made a couple of the unit tests for the `LoadingService` async because one was intermittently failing and it was apparent that the other was suffering from the same symptom, even though it was consistently passing.

Basically, `window.performance.now()` returns a time accurate to 1ms and it was entirely possible for `loadingService.register()` to run several times within the same millisecond... and when this happened, the unit test would fail.