Added the activity socket service
And the various bits and pieces to go around the service. This was essentially to prove the mechanism and required switching to a different socket.io client library than the one that had been done in isolation, due primarily to the need to pass through user data at the point of connecting.