Activity tracker
EUI-4129: Refactored the `ActivitySocketService`, along with appropriate tests. This also required some refactoring to the `ActivityService`. "socket" mode now forces websockets, rather than upgrading to them or falling back to long-polling.

EUI-4149: Made the functionality dependent on the feature flag for the Activity Tracker mode.

EUI-4126: Functionality implemented for calling the socket service when viewing a case.

EUI-4132: Functionality implemented for calling the socket service when editing a case.

EUI-4136: Functionality implemented for calling the socket service when watching cases in the list.

EUI-4146, EUI-4147: New banner to notify that a case is being watched or edited via the socket service. The new banner is actually generic and should work for both. The old one needs to be removed as it is now redundant but it needs some unit tests added for the new one first.

EUI-4148: Icons are now showing in the case list using the socket service. Again, this is a generic component that should work for the old mechanism too.